reduce-font-awesome
Currently we load the whole font-awesome icon font with 519 different
icons. We also host it ourselves, because of privacy reasons. Therefore
it is unlikely that the user will have a cached version already loaded
in his browser. So the whole font needs to be loaded upfront and thats
huge, especially because we only need a few icons. Also the use of the
fa-awesome identifier causes some troubles if another plugin uses a
font-awesome subset (e.g. event-on calendar plugin). This request
provides an own font subset only including the icons we really need
(plus flattr and VK for future implementation). By this we can reduce
the traffic caused by shariff by a lot:

Fonts: 23 KB (was 625 KB)
CSS: 9 KB (was 37 KB)

Overall we save 630 KB we really don’t need (mobile users will be
thankful) and we prevent compatibility issues with other plugins and
themes using different fa-sets.